Moved generate_docblock and associated functions to utils
